Changed the spelling correction generator to use the TextBlob library, which is much faster. The results are not perfect, but it should be fine for a quick and free service.
