+++ generators/__init__.py
@@ -3,7 +3,7 @@ from .python_functions import (
     language_translator,
     html_cleanser,
     phonetic_soundex,
-    spelling_correction,
+    textblob_spelling_correction,
     smalltalk_truncation,
     hamming_distance,
     levenshtein_distance,
@@ -35,7 +35,7 @@ for module in [
     microsoft_translator,
     deepl_translator,
     phonetic_soundex,
-    spelling_correction,
+    textblob_spelling_correction,
     smalltalk_truncation,
     hamming_distance,
     levenshtein_distance,
