+++ generators/python_functions/textblob_spelling_correction/config.py
@@ -1,11 +1,11 @@
 from util.configs import build_generator_function_config
 from util.enums import State
-from . import spelling_correction, INPUT_EXAMPLE
+from . import textblob_spelling_correction, INPUT_EXAMPLE
 
 
 def get_config():
     return build_generator_function_config(
-        function=spelling_correction,
+        function=textblob_spelling_correction,
         input_example=INPUT_EXAMPLE,
         data_type="text",
         issue_id=111,
